[WIP] Continued working on fixing client-server communication. + Clients now send a packet to be "accepted" by the server. + Server sends back the opponents to every single client once they're ready. ? Still some issues regarding inputs and states. Need further investigation.
+ Players now spawn in a circle instead of in "split lines". This allows better balancing for 3+ players.

+++ communication/State.cpp
@@ -134,7 +134,7 @@ sf::Packet& operator>>(sf::Packet &packet, State& state) {
     Position pos;
     float radius;
     bool attack;
-    int wpnId;
+    short wpnId;
     int size;
     int point;
     int timestamp;
